+++ packages/react/src/components/accordion/accordion.stories.tsx
@@ -4,7 +4,15 @@ import { AccordionArgs, accordionMeta, accordionStories } from "dso-toolkit";
 import readme from "@dso-toolkit/react/src/components/accordion/readme.md?raw";
 
 import { templateContainer } from "../../templates";
-import { addonsSections, alignmentSections, anchorSections, basicSections, subSections } from "./accordion.content";
+import {
+  activatableSections,
+  addonsSections,
+  alignmentSections,
+  anchorSections,
+  basicSections,
+  renvooiSections,
+  subSections,
+} from "./accordion.content";
 
 const meta: Meta<AccordionArgs> = {
   ...accordionMeta({ readme }),
@@ -23,6 +31,8 @@ const {
   Nested,
   AddonsSections,
   AlignmentSections,
+  Activatable,
+  RenvooiSections,
 } = accordionStories({
   templateContainer,
   storyTemplates: (templates) => {
@@ -39,7 +49,8 @@ const {
       compactBlackSections: basicSections,
       neutralSections: basicSections,
       subSections: subSections(templates),
-      renvooiSections: subSections(templates),
+      renvooiSections,
+      activatableSections,
     };
   },
 });
@@ -54,4 +65,6 @@ export {
   Nested,
   AddonsSections,
   AlignmentSections,
+  RenvooiSections,
+  Activatable,
 };