+++ storybook/src/components/accordion/accordion.css-stories.ts
@@ -4,6 +4,7 @@ import { AccordionArgs, accordionMeta, accordionStories } from "dso-toolkit";
 import { templateContainer } from "../../templates";
 
 import {
+  activatableSections,
   addonsSections,
   alignmentSections,
   anchorSections,
@@ -20,26 +21,48 @@ const meta: Meta<AccordionArgs> = {
 
 export default meta;
 
-const { Default, Compact, Conclusion, HandleAnchors, Nested, AddonsSections, AlignmentSections, RenvooiSections } =
-  accordionStories({
-    templateContainer,
-    storyTemplates: (templates) => {
-      const { accordionTemplate } = templates;
+const {
+  Default,
+  Compact,
+  CompactBlack,
+  Conclusion,
+  HandleAnchors,
+  Nested,
+  AddonsSections,
+  AlignmentSections,
+  RenvooiSections,
+  Activatable,
+} = accordionStories({
+  templateContainer,
+  storyTemplates: (templates) => {
+    const { accordionTemplate } = templates;
 
-      return {
-        accordionTemplate,
-        basicSections: basicSections(templates),
-        addonsSections,
-        alignmentSections,
-        anchorSections: anchorSections(templates),
-        conclusionSections: basicSections(templates),
-        compactSections: basicSections(templates),
-        compactBlackSections: basicSections(templates),
-        neutralSections: basicSections(templates),
-        subSections: subSections(templates),
-        renvooiSections,
-      };
-    },
-  });
+    return {
+      accordionTemplate,
+      basicSections: basicSections(templates),
+      addonsSections,
+      alignmentSections,
+      anchorSections: anchorSections(templates),
+      conclusionSections: basicSections(templates),
+      compactSections: basicSections(templates),
+      compactBlackSections: basicSections(templates),
+      neutralSections: basicSections(templates),
+      subSections: subSections(templates),
+      renvooiSections,
+      activatableSections,
+    };
+  },
+});
 
-export { Default, Compact, Conclusion, HandleAnchors, Nested, AddonsSections, AlignmentSections, RenvooiSections };
+export {
+  Default,
+  Compact,
+  CompactBlack,
+  Conclusion,
+  HandleAnchors,
+  Nested,
+  AddonsSections,
+  AlignmentSections,
+  RenvooiSections,
+  Activatable,
+};