+++ storybook/src/components/accordion/accordion.css-template.ts
@@ -7,7 +7,7 @@ import { ComponentImplementation } from "../../templates";
 export const cssAccordion: ComponentImplementation<Accordion<TemplateResult>> = {
   component: "accordion",
   implementation: "html-css",
-  template: ({ attachmentsCounterTemplate, iconTemplate, labelTemplate, renvooiTemplate }) =>
+  template: ({ attachmentsCounterTemplate, iconTemplate, labelTemplate, renvooiTemplate, slideToggleTemplate }) =>
     function accordionTemplate(accordion) {
       const statusMap = new Map<string, string>([
         ["success", "succes"],
@@ -43,12 +43,19 @@ export const cssAccordion: ComponentImplementation<Accordion<TemplateResult>> = 
       function accordionHandleContent(accordion: Accordion<TemplateResult>, section: AccordionSection<TemplateResult>) {
         const ariaExpanded = (section.open ?? false).toString();
         const children = accordionHandleChildren(accordion, section);
+        const showSlideToggle = section.activatable && accordion.variant === "compact-black" && !accordion.reverseAlign;
+        const slideToggle = showSlideToggle
+          ? slideToggleTemplate({
+              accessibleLabel: "Toon op kaart",
+              checked: !!section.active,
+            })
+          : nothing;
 
         if (section.handleUrl) {
-          return html`<a href=${section.handleUrl} aria-expanded=${ariaExpanded}>${children}</a>`;
+          return html`<a href=${section.handleUrl} aria-expanded=${ariaExpanded}>${children}</a>${slideToggle}`;
         }
 
-        return html`<button type="button" aria-expanded=${ariaExpanded}>${children}</button>`;
+        return html`<button type="button" aria-expanded=${ariaExpanded}>${children}</button>${slideToggle}`;
       }
 
       function accordionHandleTemplate(
@@ -74,6 +81,7 @@ export const cssAccordion: ComponentImplementation<Accordion<TemplateResult>> = 
         section: AccordionSection<TemplateResult>,
       ): TemplateResult {
         const hasNestedAccordion = section.content?.strings.includes("<dso-accordion") ?? false;
+        const showSlideToggle = section.activatable && accordion.variant === "compact-black" && !accordion.reverseAlign;
 
         if (!section.heading) {
           section.heading = "h2";
@@ -86,6 +94,7 @@ export const cssAccordion: ComponentImplementation<Accordion<TemplateResult>> = 
               "dso-open": !!section.open,
               "dso-nested-accordion": hasNestedAccordion,
               [`dso-accordion-wijzigactie-${section.wijzigactie}`]: !!section.wijzigactie,
+              "dso-accordion-section-activate": !!showSlideToggle,
             })}"
           >
             ${accordionHandleTemplate(accordion, section)}